requirements.txt
confluent-kafka==2.5.3
authlib
httpx
python-dotenv
fastavro
jsonschema
orjson
requests